requirements.txt
# API and Web
fastapi>=0.95.1
uvicorn>=0.22.0
streamlit>=1.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
starlette>=0.27.0
aiohttp>=3.8.4

# Data Processing
pandas>=1.5.3
numpy>=1.24.3

# Image Processing
pillow>=9.5.0

# Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# AI Integration
openai>=1.0.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0